requirements.txt
selenium==4.1.2
gppt==2.1.2
tweepy==4.10.0
flask-openapi3==2.0.0
python-dotenv==0.20.0
nanoid==2.0.0
gunicorn==20.1.0
